+++ src/client/routes/_layout.collection.$collectionId.tsx
@@ -2,6 +2,8 @@ import { createFileRoute, redirect } from '@tanstack/react-router';
 import { queryOptions } from '@tanstack/react-query';
 import { api, bidirectionalInfiniteQuery, queryClient } from '../online/api';
 import toast from 'react-hot-toast';
+import { z } from 'zod';
+import { zodValidator } from '@tanstack/zod-adapter';
 
 export const collectionQueryOptions = (collectionId: string) =>
   queryOptions({
@@ -9,23 +11,38 @@ export const collectionQueryOptions = (collectionId: string) =>
     queryFn: () => api.getCollectionBrief(collectionId),
   });
 
-export const collectionInfiniteQueryOptions = (collectionId: string) =>
+export const collectionInfiniteQueryOptions = (
+  collectionId: string,
+  sort?: 'asc' | 'desc'
+) =>
   bidirectionalInfiniteQuery(
-    ['collection', collectionId, 'puzzles'],
+    ['collection', collectionId, 'puzzles', sort],
     (cursorBefore, cursorAfter) =>
-      api.listCollectionPuzzles(collectionId, cursorBefore, cursorAfter)
+      api.listCollectionPuzzles(collectionId, sort, cursorBefore, cursorAfter)
   );
 
+export const collectionSortSchema = z.object({
+  sort: z.enum(['asc', 'desc']).optional().catch(undefined),
+});
+
 export const Route = createFileRoute('/_layout/collection/$collectionId')({
+  validateSearch: zodValidator(collectionSortSchema),
   remountDeps: ({ params }) => params.collectionId,
-  loader: async ({ params }) => {
+  loaderDeps: ({ search: { sort } }) => ({ sort }),
+  loader: async ({
+    params,
+    deps: { sort },
+  }: {
+    params: { collectionId: string };
+    deps: { sort?: 'asc' | 'desc' };
+  }) => {
     try {
       await queryClient.ensureQueryData(
         collectionQueryOptions(params.collectionId)
       );
       // we can render the page and suspend while waiting for this
       void queryClient.ensureInfiniteQueryData(
-        collectionInfiniteQueryOptions(params.collectionId)
+        collectionInfiniteQueryOptions(params.collectionId, sort)
       );
     } catch (error) {
       toast.error((error as Error).message);